script/requirements.txt
google-api-core==2.2.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-monitoring==2.8.0
googleapis-common-protos==1.53.0
py-zabbix==1.1.7
